src/app/create-event/create-event.component.spec.ts
import { ComponentFixture, TestBed } from '@angular/core/testing';

import { CreateEventComponent } from './create-event.component';
import {HttpClientModule} from "@angular/common/http";
import {RouterTestingModule} from "@angular/router/testing";
import {FormBuilder, FormsModule, ReactiveFormsModule} from "@angular/forms";
import {HttpClientTestingModule, HttpTestingController} from "@angular/common/http/testing";
import {formatDate} from "@angular/common";

describe('CreateEventComponent', () => {
  let component: CreateEventComponent;
  let fixture: ComponentFixture<CreateEventComponent>;
  let httpTestingController: HttpTestingController;

  beforeEach(async () => {
    await TestBed.configureTestingModule({
      imports: [HttpClientModule, RouterTestingModule, ReactiveFormsModule, FormsModule, HttpClientTestingModule],
      declarations: [ CreateEventComponent ]
    })
    .compileComponents();
  });

  beforeEach(() => {
    fixture = TestBed.createComponent(CreateEventComponent);
    component = fixture.componentInstance;
    fixture.detectChanges();

    httpTestingController = TestBed.inject(HttpTestingController);
  });

  it('should create', () => {
    expect(component).toBeTruthy();
  });

  it('should make post request', function () {
    component.submitHandler();
    const req = httpTestingController.expectOne('http://127.0.0.1:5000/events');
    expect(req.request.method).toEqual('POST');
  });

  it('should make form', function () {
    const formBuilder = new FormBuilder();
    const form = formBuilder.group({
      title: '',
      date: ''
    });
    component.ngOnInit();
    expect(component.form.getRawValue()).toEqual(form.getRawValue());
  });
});
